refactoring profile page settings (no pain, reusable)
