+++ webapp/src/components/atoms/TextField/NumberField.tsx
@@ -1,0 +1,5 @@
+import { TextField, TextFieldProps } from "@mui/material";
+
+export default function NumberField(props: TextFieldProps) {
+  return <TextField {...props} type="number" fullWidth size="small" />;
+}
